requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.6
django-webpush==0.3.3
http-ece==1.1.0
idna==2.10
py-vapid==1.7.1
pycparser==2.20
pytz==2021.1
pywebpush==1.9.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
